Add roles section tests to UserManagementPage for role rendering and creation
